requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
scikit-learn>=1.3.0
motor>=3.3.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pymongo>=4.5.0
numpy>=1.24.0
joblib>=1.3.0